+++ com.unity.render-pipelines.core/Editor/Deprecated.cs
@@ -1,0 +1,31 @@
+using System;
+
+namespace UnityEditor.Rendering
+{
+    /// <summary>
+    /// This attributes tells a <see cref="VolumeComponentEditor"/> class which type of
+    /// <see cref="VolumeComponent"/> it's an editor for.
+    /// When you make a custom editor for a component, you need put this attribute on the editor
+    /// class.
+    /// </summary>
+    /// <seealso cref="VolumeComponentEditor"/>
+    [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
+    [Obsolete("VolumeComponentEditor property has been deprecated. Please use CustomEditor.")]
+    public sealed class VolumeComponentEditorAttribute : CustomEditor
+    {
+        /// <summary>
+        /// A type derived from <see cref="VolumeComponent"/>.
+        /// </summary>
+        public readonly Type componentType;
+
+        /// <summary>
+        /// Creates a new <see cref="VolumeComponentEditorAttribute"/> instance.
+        /// </summary>
+        /// <param name="componentType">A type derived from <see cref="VolumeComponent"/></param>
+        public VolumeComponentEditorAttribute(Type componentType)
+            : base(componentType, true)
+        {
+            this.componentType = componentType;
+        }
+    }
+}