+++ com.unity.render-pipelines.core/Editor/Volume/VolumeComponentListEditor.cs
@@ -60,17 +60,7 @@ namespace UnityEditor.Rendering
         SerializedObject m_SerializedObject;
         SerializedProperty m_ComponentsProperty;
 
-        Dictionary<Type, Type> m_EditorTypes; // Component type => Editor type
-        List<VolumeComponentEditor> m_Editors;
-
-        static Dictionary<Type, string> m_EditorDocumentationURLs;
-
-        int m_CurrentHashCode;
-
-        static VolumeComponentListEditor()
-        {
-            ReloadDocumentation();
-        }
+        List<VolumeComponentEditor> m_Editors = new List<VolumeComponentEditor>();
 
         /// <summary>
         /// Creates a new instance of <see cref="VolumeComponentListEditor"/> to use in an
@@ -96,22 +86,8 @@ namespace UnityEditor.Rendering
 
             this.asset = asset;
             m_SerializedObject = serializedObject;
-            m_ComponentsProperty = serializedObject.Find((VolumeProfile x) => x.components);
-            Assert.IsNotNull(m_ComponentsProperty);
 
-            m_EditorTypes = new Dictionary<Type, Type>();
-            m_Editors = new List<VolumeComponentEditor>();
-
-            foreach (var editorType in TypeCache.GetTypesWithAttribute<VolumeComponentEditorAttribute>())
-            {
-                var attribute = (VolumeComponentEditorAttribute)editorType.GetCustomAttributes(typeof(VolumeComponentEditorAttribute), false)[0];
-                m_EditorTypes.Add(attribute.componentType, editorType);
-            }
-
-            // Create editors for existing components
-            var components = asset.components;
-            for (int i = 0; i < components.Count; i++)
-                CreateEditor(components[i], m_ComponentsProperty.GetArrayElementAtIndex(i));
+            RefreshEditors();
 
             // Keep track of undo/redo to redraw the inspector when that happens
             Undo.undoRedoPerformed += OnUndoRedoPerformed;
@@ -152,12 +128,10 @@ namespace UnityEditor.Rendering
                 }
             }
 
-            if (!m_EditorTypes.TryGetValue(componentType, out Type editorType))
-                editorType = typeof(VolumeComponentEditor);
-
-            var editor = (VolumeComponentEditor)Editor.CreateEditor(component); //(VolumeComponentEditor)Activator.CreateInstance(editorType);
-            //editor.Init(component, m_BaseEditor);
+            var editor = (VolumeComponentEditor)Editor.CreateEditor(component);
+            editor.inspector = m_BaseEditor;
             editor.baseProperty = property.Copy();
+            editor.Init();
 
             if (forceOpen)
                 editor.baseProperty.isExpanded = true;
@@ -168,14 +142,18 @@ namespace UnityEditor.Rendering
                 m_Editors[index] = editor;
         }
 
+        int m_CurrentHashCode;
         void RefreshEditors()
         {
-            // Disable all editors first
-            foreach (var editor in m_Editors)
-                editor.OnDisable();
+            if (m_Editors.Count > 0)
+            {
+                // Disable all editors first
+                foreach (var editor in m_Editors)
+                    editor.OnDisable();
 
-            // Remove them
-            m_Editors.Clear();
+                // Remove them
+                m_Editors.Clear();
+            }
 
             // Refresh the ref to the serialized components in case the asset got swapped or another
             // script is editing it while it's active in the inspector
@@ -187,6 +165,8 @@ namespace UnityEditor.Rendering
             var components = asset.components;
             for (int i = 0; i < components.Count; i++)
                 CreateEditor(components[i], m_ComponentsProperty.GetArrayElementAtIndex(i));
+
+            m_CurrentHashCode = asset.GetComponentListHashCode();
         }
 
         /// <summary>
@@ -202,7 +182,6 @@ namespace UnityEditor.Rendering
                 editor.OnDisable();
 
             m_Editors.Clear();
-            m_EditorTypes.Clear();
 
             // ReSharper disable once DelegateSubtraction
             Undo.undoRedoPerformed -= OnUndoRedoPerformed;
@@ -221,7 +200,6 @@ namespace UnityEditor.Rendering
             if (asset.isDirty || asset.GetComponentListHashCode() != m_CurrentHashCode)
             {
                 RefreshEditors();
-                m_CurrentHashCode = asset.GetComponentListHashCode();
                 asset.isDirty = false;
             }
 
@@ -230,14 +208,11 @@ namespace UnityEditor.Rendering
 
             using (new EditorGUI.DisabledScope(!isEditable))
             {
-                // Component list
                 for (int i = 0; i < m_Editors.Count; i++)
                 {
                     var editor = m_Editors[i];
                     var title = editor.GetDisplayTitle();
                     int id = i; // Needed for closure capture below
-
-                    m_EditorDocumentationURLs.TryGetValue(editor.target.GetType(), out var documentationURL);
 
                     CoreEditorUtils.DrawSplitter();
                     bool displayContent = CoreEditorUtils.DrawHeaderToggle(
@@ -247,7 +222,7 @@ namespace UnityEditor.Rendering
                         pos => OnContextClick(pos, editor, id),
                         editor.hasAdditionalProperties ? () => editor.showAdditionalProperties : (Func<bool>)null,
                         () => editor.showAdditionalProperties ^= true,
-                        documentationURL
+                        Help.GetHelpURLForObject(editor.volumeComponent)
                     );
 
                     if (displayContent)
@@ -523,33 +498,6 @@ namespace UnityEditor.Rendering
             string typeData = clipboard.Substring(clipboard.IndexOf('|') + 1);
             Undo.RecordObject(targetComponent, "Paste Settings");
             JsonUtility.FromJsonOverwrite(typeData, targetComponent);
-        }
-
-        static void ReloadDocumentation()
-        {
-            if (m_EditorDocumentationURLs == null)
-                m_EditorDocumentationURLs = new Dictionary<Type, string>();
-            m_EditorDocumentationURLs.Clear();
-
-            string GetVolumeComponentDocumentation(Type component)
-            {
-                var attrs = component.GetCustomAttributes(false);
-                foreach (var attr in attrs)
-                {
-                    if (attr is HelpURLAttribute attrDocumentation)
-                        return attrDocumentation.URL;
-                }
-
-                // There is no documentation for this volume component.
-                return null;
-            }
-
-            var componentTypes = CoreUtils.GetAllTypesDerivedFrom<VolumeComponent>();
-            foreach (var componentType in componentTypes)
-            {
-                if (!m_EditorDocumentationURLs.ContainsKey(componentType))
-                    m_EditorDocumentationURLs.Add(componentType, GetVolumeComponentDocumentation(componentType));
-            }
         }
     }
 }