+++ com.unity.render-pipelines.core/CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Fixed issue when changing volume profiles at runtime with a script (case 1364256).
 
+### Changed
+- Volume Component editor are now specified by `CustomEditorAttribute` instead of `VolumeComponentEditorAttribute`. 
+
 ## [13.1.1] - 2021-10-04
 
 ### Added